# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TRPG-Simulation API
  description: |
    線上 TRPG 輔助平台後端 API

    最近更新（v0.12.0, 2026-03-03）:
    - 背包系統：give_item / remove_item WebSocket actions
    - 投票系統：player_votes 事件
    - 即時轉場更新：transitions_updated 事件
    - GM 變數修改：set_variable WebSocket action
    - DELETE /api/v1/sessions/{id} 場次刪除
    完整變更紀錄見 docs/api-changelog.md
  version: 0.12.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/sessions/{id}:
    get:
      operationId: getSession
      summary: 取得場次
      description: 依 ID 取得場次詳情。僅 GM 或場次成員可存取。
      tags:
        - sessions
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: 場次 UUID
      responses:
        "200":
          description: 場次詳情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          description: 未提供有效的 access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: 無權限存取此場次
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 場次不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteSession
      summary: 刪除場次
      description: 刪除一個處於 lobby 狀態的場次。僅 GM 可操作。
      tags:
        - sessions
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: 場次 UUID
      responses:
        "204":
          description: 場次已刪除
        "401":
          description: 未提供有效的 access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: 非場次 GM
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 場次不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: 場次非 lobby 狀態，無法刪除
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token。在 Authorization header 中以 Bearer scheme 傳送。
        Token 有效期 15 分鐘，過期後需透過 /api/v1/auth/refresh 取得新 token。
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: 機器可讀的錯誤代碼（UPPER_SNAKE_CASE）
          example: VALIDATION_ERROR
        message:
          type: string
          description: 人類可讀的錯誤說明
          example: Email format is invalid
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
          description: 欄位級別的錯誤資訊（可選）
    ErrorDetail:
      type: object
      required:
        - field
        - reason
      properties:
        field:
          type: string
          description: 發生錯誤的欄位名稱
          example: email
        reason:
          type: string
          description: 錯誤原因
          example: must be a valid email address
...
    SessionResponse:
      type: object
      required:
        - id
        - scenarioId
        - gmId
        - status
        - inviteCode
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: 場次 UUID
          example: 770e8400-e29b-41d4-a716-446655440000
        scenarioId:
          type: string
          format: uuid
          description: 劇本 UUID
          example: 550e8400-e29b-41d4-a716-446655440000
        gmId:
          type: string
          format: uuid
          description: GM 用戶 UUID
          example: 660e8400-e29b-41d4-a716-446655440000
        status:
          type: string
          enum:
            - lobby
            - active
            - paused
            - completed
            - abandoned
          description: 場次狀態
          example: lobby
        inviteCode:
          type: string
          description: 邀請碼（6 位大寫英數）
          example: ABC123
        createdAt:
          type: string
          format: date-time
          description: 建立時間（RFC 3339）
          example: "2026-02-27T12:00:00Z"
        startedAt:
          type: string
          format: date-time
          description: 遊戲開始時間（可為 null）
          example: "2026-02-27T12:30:00Z"
        endedAt:
          type: string
          format: date-time
          description: 遊戲結束時間（可為 null）
          example: "2026-02-27T14:00:00Z"
...